removes int size from mem index type

+++ waku/v2/protocol/waku_rln_relay/waku_rln_relay_types.nim
@@ -54,7 +54,7 @@ type RateLimitProof* = object
   ## see details in https://hackmd.io/tMTLMYmTR5eynw2lwK9n1w?view#Nullifiers
   nullifier*: Nullifier
   
-type MembershipIndex* = uint64
+type MembershipIndex* = uint
 
 type WakuRLNRelay* = ref object 
   membershipKeyPair*: MembershipKeyPair
